AgenteMaestro/pyproject.toml
[project]
name = "agente-maestro"
version = "1.0.0"
description = "AgenteMaestro - Agente autônomo com auto-melhoria"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.6.3",
    "chromadb>=1.0.13",
    "fastapi>=0.115.13",
    "groq>=0.28.0",
    "openai>=1.88.0",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "tavily-python>=0.7.7",
    "uvicorn>=0.34.3",
    "yfinance>=0.2.63",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "psutil>=5.9.0",
    "black>=24.0.0",
    "watchdog>=4.0.0",
]
